Add countdown to js bundle, remove add_script()
Since the addition of webpack we're making a JavaScript bundle that
contains essentially everything; with this change, we're also
including the countdown JavaScript in the bundle.

Since all the JavaScript code is in the JS bundle, there's no need
to have scripts added dynamically by PHP code. The PHP code should
always include the bundle, and nothing else, so this patch removes
add_script() and add_scripts() functions.