Add setup for system tests with docker compose
In order to be able to run possible future automated system tests, as
well as being able to play around with the system while developing,
create some system test setup files. You can bring the system online
using
  docker compose \
         --project-directory . \
         --file tests/system_tests/setup/docker-compose.yaml \
         up
and take it back down with the "docker compose down --volumes" (and
the same project directory and file options). Login credentials are in
the "secrets" Docker volume.