Map all JSON URLs to new api.php, not exportcards
Create a new api.php file with an embryo of a REST API, and map
all URLs ending with .json to api.php. The only such URL we have
at the moment is the export cards functionality, so we remove
exportcards.php and let api.php handle requests for cards.json.